pyproject.toml
[tool.poetry]
name = "disbot"
version = "0.1.0"
description = "A Discord bot to help in server maintenance."
authors = ["Jatin Dua <jdua004@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = {extras = ["voice"], version = "^2.2.3"}
python-dotenv = "^1.0.0"
asyncio = "^3.4.3"
pylint = "^2.17.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
